+++ apps/user-manager/src/main/java/com/warlock/user/controller/UserController.java
@@ -2,11 +2,13 @@ package com.warlock.user.controller;
 
 import com.warlock.user.model.LoginRequest;
 import com.warlock.user.model.UserToken;
+import com.warlock.user.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -20,6 +22,13 @@ import static org.springframework.http.ResponseEntity.ok;
 @Controller
 @Slf4j
 public class UserController {
+    final UserService userService;
+
+    @Autowired
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
+
     @PostMapping(value = "/login", produces = MediaType.APPLICATION_JSON_VALUE)
     @Operation(
             summary = "Logs in a user",
@@ -37,9 +46,10 @@ public class UserController {
             description = "User not recognized",
             content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE, schema = @Schema(hidden = true)))
     public ResponseEntity<UserToken> login(
-            @RequestBody @Valid LoginRequest request
+            @RequestBody @Valid LoginRequest loginRequest
     ) {
-        log.info("Request username: {} | Request password: {}", request.getUsername(), request.getPassword());
-        return ok(UserToken.builder().token("Suck it").build());
+        log.info("Controller login called");
+        var token = userService.login(loginRequest);
+        return ok(UserToken.builder().token(token).build());
     }
 }